refactor: isBurst() -> carDraw()로 메서드명 변경
