+++ src/main/java/blackjack/domain/gamer/Gamer.java
@@ -8,7 +8,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public abstract class Gamer {
-    public static final int INIT_DISTRIBUTION_COUNT = 2;
     public static final int MAX_CARD_VALUE = 21;
 
     private final Name name;
@@ -61,5 +60,5 @@ public abstract class Gamer {
         return cards.size();
     }
 
-    abstract boolean isBurst();
+    abstract boolean canDraw();
 }
