+++ src/test/java/blackjack/domain/gamer/DealerTest.java
@@ -20,6 +20,6 @@ class DealerTest {
         dealer.addCard(card1);
         dealer.addCard(card2);
 
-        assertThat(dealer.isBurst()).isEqualTo(true);
+        assertThat(dealer.canDraw()).isEqualTo(true);
     }
 }
